chore: nuke GitExt trait, inline into TestContext
Removes the GitExt trait and git.rs module entirely. All methods
(init_git, remote, local, consensus) are now inherent methods on
TestContext in common/mod.rs. Also removes dead set_timestamps fn.

Closes #146
